refactor(cost-tracker): consolidate session-loader into _sessions.mjs (iter 73)
Iter 68 consolidated PRICING. This iter applies the same DRY pattern to
session loading — the single largest remaining source of duplication across
the analytics scripts.

Seven scripts (anomaly, burn, projection, counterfactual, conversation,
budget, summary) each carried byte-identical copies of:
  - memoryListSessionKeys / memoryListSessionRecords  (~15 LOC each)
  - memoryRetrieve                                    (~10 LOC each)
  - parseDurationMs                                   (~6 LOC each)
  - sessionTs                                         (~4 LOC each)

Now consolidated in scripts/_sessions.mjs:
  loadSessions(ns)           → array of session records
  memoryListSessionKeys(ns)  → keys starting with `session-`
  memoryListAllKeys(ns)      → ALL keys (for budget-config / fed-spend lookups)
  memoryRetrieve(ns, key)    → parsed JSON value or null
  sessionTs(rec)             → ms-since-epoch from capturedAt|endedAt|startedAt
  parseDurationMs(spec)      → ms for Nh/Nd/Nw/Nm; null on parse fail

NET: −163 LOC across 7 scripts.

MATH BYTE-IDENTICAL — verified end-to-end with fixture (5 baseline sessions
$0.08-$0.12 + 1 outlier $5.00):
  - anomaly: median $0.10, modified-z 330.505 (same as pre-refactor)
  - burn: deltaPct 5163.16 (same as pre-refactor)
  - exit codes preserved (anomaly@1: 1, burn@50: 1, health: 1)

SMOKE UPDATED
- Steps 39d (burn) and 39f (anomaly) now accept "spawnSync OR _sessions.mjs"
  for the safe-shell-out invariant (consolidation moved spawnSync into the
  vetted helper).
- New step 42b: dedicated invariant for _sessions.mjs itself. Asserts
  spawnSync usage + all 6 exports + that every consumer script actually
  imports from it.

Smoke contract: 55 → 56 passing. Version 0.21.0 → 0.21.1 (patch — internal
refactor with no behavior change).
