+++ plugins/ruflo-cost-tracker/scripts/budget.mjs
@@ -13,6 +13,10 @@
 //   BUDGET_QUIET=1            machine-readable JSON only
 
 import { spawnSync } from 'node:child_process';
+// iter 73 — shared session-loader (was duplicated across 6 scripts).
+// Only the session-list path consolidates; budget-config reads stay local
+// because they have their own "latest stamp" upsert resolution logic.
+import { loadSessions } from './_sessions.mjs';
 
 // ADR-100 / #1748 Issue 3 — opt into cli-core's lite path with CLI_CORE=1.
 // Cold-cache wall-time drops from ~25s to ~2s. JSON backend instead of
@@ -87,29 +91,6 @@ function memoryRetrieve(key) {
   return memoryRetrieveOne(key);
 }
 
-function memoryListSessionRecords() {
-  // Use --format json so keys aren't truncated with `...` in tabular output.
-  const r = spawnSync('npx', [
-    CLI_PKG, 'memory', 'list',
-    '--namespace', NS, '--format', 'json',
-  ], { stdio: ['ignore', 'pipe', 'pipe'], encoding: 'utf-8', shell: process.platform === 'win32' });
-  if (r.status !== 0) return [];
-  // CLI may emit a small banner before the JSON array; extract the first '['..']' block.
-  const m = /\[[\s\S]*\]/.exec(r.stdout || '');
-  if (!m) return [];
-  let entries;
-  try { entries = JSON.parse(m[0]); } catch { return []; }
-  const sessionKeys = entries
-    .map((e) => e.key)
-    .filter((k) => typeof k === 'string' && k.startsWith('session-'));
-  const records = [];
-  for (const k of sessionKeys) {
-    const rec = memoryRetrieve(k);
-    if (rec) records.push(rec);
-  }
-  return records;
-}
-
 function periodFilter(period) {
   const now = Date.now();
   const day = 24 * 3600 * 1000;
@@ -175,7 +156,7 @@ function cmdGet() {
 function cmdCheck() {
   const cfg = memoryRetrieve(KEY);
   const period = process.env.BUDGET_PERIOD || 'all';
-  const records = memoryListSessionRecords();
+  const records = loadSessions(NS);
   const filt = periodFilter(period);
   const filtered = records.filter((r) => filt(r.capturedAt || r.endedAt));
   const totalSpend = filtered.reduce((s, r) => s + (r.total_cost_usd || 0), 0);
